0.2.6 - Fixing PyPI setup
My setup.py was completely broken and wrong because the requirements.txt
file was never included in the PyPI package.  This meant that it was
really only possible to install the package by cloning the repo and
running setup.py from here, where requirements.txt would be available.

Should be fixed now.